api/requirements.txt
flask
Flask-SQLAlchemy
flask-marshmallow
flask-smorest
marshmallow-sqlalchemy
stripe
psycopg2